Move submodule to new repo
So that I have a "clean" fork of pelican-bootstrap3 and a modified fork
that I work on. (If I wanted to say pull request to the main
pelican-bootstrap3 repo, it would be a mess otherwise.)

Due to version of git on my machine being too old to use a branch
of a submodule.