Update pelican-bootstrap3-sm theme to latest
Following StevenMaude/pelican-bootstrap3-sm#12.

This also changes the theme location.

Due to the theme moving to the getpelican/pelican-themes repository, and
becoming a subdirectory, it was easier to get commits to apply via
`git cherry-pick` when the structure was the same. The result of this is
that the path has changed here too.