Use i18n_subsites plugin
This change is in preparation for updating the theme.

i18n support was added for the theme which is fine,
except that this is not optional and breaks the theme if you don't have
this plugin enabled.

This is a little bit annoying having a dependency just to do nothing.

But, see getpelican/pelican-themes#465; the only other options are
hacks, which may break in future. I'd prefer to run things as
sanctioned, avoiding any unexpected breakage in future.