Reduce the number of tag cloud steps
Updating the tag cloud plugin to the latest commit now results in some
text in tags going very small. This could be fixed by altering the
style.css, but I'd prefer not to make changes there unless necessary
(because it can lead to merge conflicts in future) or override it in
custom.css (as there are extra rules).

Limiting the steps isn't a bad thing anyway as it restricts the tag
classes to fewer possible classes.