Comment out tags menuitem in config
The problem is that this is a relative link, meaning that it is broken
when clicking on "Tags" when visiting an article page.

Don't think there is a way to specify a relative link here, see
https://github.com/getpelican/pelican/issues/2272

The easiest fix is to remove the link, and reduces clutter on the page.